common_table_website/requirements.txt
asgiref==3.7.2
Django==3.2.24
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.4.0
